refactor(worker): extract agent-turn stream engine
Move prepare/runStreamAttempt, open-suffix resume, and the compaction-retry
loop into runTurnStreamAttempt. Share first-request prep stamps on eventing
so Codex diagnostic callbacks still see stream mutations.
